requirements.txt
google-auth==2.28.1
google-auth-httplib2==0.2.0
fastapi==0.115.5
uvicorn[standard]==0.32.0
motor==3.5.1
pymongo==4.8.0
pydantic==2.9.2
pydantic[email]==2.9.2
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
cloudinary==1.41.0
requests